Add Archivo Black and MailChimp Waiting List form
I used Archivo Black as another font when designing
the initial logo. I added this here to add contrast
to the headlines and to the "Goal" of LifeGoals.
I think it looks better that way.

I removed the download buttons and put a MailChimp
Waiting List form because the app is not yet ready
for download to the public.

I also commented inthe social icons which I will
comment out once we have social network accounts.
I also adjusted the position of the screenshots a
bit to align them. I think it looks better now.

Signed-off-by: Jj <johnjullies@gmail.com>